Move constants to constants.rs
